CU-168 - Do not validate at model.save() level
In Django docs they do not recommend to implement the validation
inside save(). The reason not to do so is because this
validation should be done right after the data is received.
Otherwise, we would be doing validation two times, costing us more
processing resources.

Changing this would not prevent anyone to introduce not clean data
directly through the Django shell, but neither would do if we
introduce data manually into the DB.

https://docs.djangoproject.com/en/2.0/ref/models/instances/#validating-objects

Doing this would prevent all 500 errors when submiting not validated
data.